monitor-status-log/requirements.txt
pymongo
kafka-python
paramiko==2.4.2
dnspython
docker
